requirements.txt
Django==5.2.5
djangorestframework==3.16.1
psycopg==3.2.9
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
